requirements.txt
pymupdf>=1.24.0
requests==2.31.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
langchain-groq
groq>=0.5.0
python-dotenv>=0.20.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5